services/realtime/requirements.txt
fastapi[standard]
sqlalchemy
asyncpg
redis
python-json-logger
sentry-sdk
PyJWT
cryptography
requests
python-decouple
